+++ src/thirdLIB/OmniFrameworks/OmniFoundation.framework/Versions/A/Headers/OFKnownKeyDictionaryTemplate.h
@@ -1,11 +1,9 @@
-// Copyright 1998-2005 Omni Development, Inc.  All rights reserved.
+// Copyright 1998-2019 Omni Development, Inc. All rights reserved.
 //
 // This software may only be used and reproduced according to the
 // terms in the file OmniSourceLicense.html, which should be
 // distributed with this project and can also be found at
 // <http://www.omnigroup.com/developer/sourcecode/sourcelicense/>.
-//
-// $Header: svn+ssh://source.omnigroup.com/Source/svn/Omni/tags/OmniSourceRelease/2008-09-09/OmniGroup/Frameworks/OmniFoundation/DataStructures.subproj/OFKnownKeyDictionaryTemplate.h 68913 2005-10-03 19:36:19Z kc $
 
 #import <OmniFoundation/OFObject.h>
 
@@ -17,9 +15,9 @@ This class holds information common to a set of OFMutableKnownKeyDictionaries.  
 */
 {
 @public // These should really only be accessed by OFMutableKnownKeyDictionary
-    NSArray       *_keyArray;
-    unsigned int   _keyCount;
-    NSObject      *_keys[0];
+    NSArray *_keyArray;
+    NSUInteger _keyCount;
+    NSObject *_keys[0];
 }
 
 + (OFKnownKeyDictionaryTemplate *) templateWithKeys: (NSArray *) keys;
